+++ src/pycram/orm/__init__.py
@@ -1,1 +1,1 @@
-from . import base, action_designator, object_designator, motion_designator, task
+from . import base, action_designator, object_designator, motion_designator, tasktree
